+++ ts_noise/src/ikpsk2.rs
@@ -6,7 +6,11 @@ use ts_keys::X25519KeyPair;
 use zerocopy::FromBytes;
 
 use crate::{
-    core::{Psk, Session, State},
+    core::{
+        Psk,
+        Role::{Initiator, Responder},
+        Session, State,
+    },
     messages::{Init, Pod, Resp},
 };
 
@@ -87,7 +91,7 @@ impl ReceivedHandshake {
             .mix_dh(&my_ephemeral.private, &self.peer_static_pub) // se
             .mix_psk(psk) // psk
             .seal(&mut [], &mut response.auth_tag) // payload
-            .finish_as_responder()
+            .finish(Responder)
     }
 }
 
@@ -175,7 +179,7 @@ impl<P: Pod> SentHandshake<P> {
             .mix_psk(psk) // psk
             .open(&mut [], &packet.auth_tag)
             .ok_or(self)?
-            .finish_as_initiator();
+            .finish(Initiator);
 
         Ok(ret)
     }
@@ -238,7 +242,15 @@ mod tests {
             panic!("initiator failed to finalize handshake");
         };
 
-        assert_eq!(init_session.send, resp_session.recv);
-        assert_eq!(init_session.recv, resp_session.send);
+        assert_eq!(
+            init_session.initiator_to_responder,
+            resp_session.initiator_to_responder
+        );
+        assert_eq!(
+            init_session.responder_to_initiator,
+            resp_session.responder_to_initiator
+        );
+        assert_eq!(init_session.role, Initiator);
+        assert_eq!(resp_session.role, Responder);
     }
 }